split to make compliance happy
