+++ Reflection_Engine/Create/Type/GenericType.cs
@@ -1,0 +1,63 @@
+﻿/*
+ * This file is part of the Buildings and Habitats object Model (BHoM)
+ * Copyright (c) 2015 - 2021, the respective contributors. All rights reserved.
+ *
+ * Each contributor holds copyright over their respective contributions.
+ * The project versioning (Git) records all such contribution source information.
+ *                                           
+ *                                                                              
+ * The BHoM is free software: you can redistribute it and/or modify         
+ * it under the terms of the GNU Lesser General Public License as published by  
+ * the Free Software Foundation, either version 3.0 of the License, or          
+ * (at your option) any later version.                                          
+ *                                                                              
+ * The BHoM is distributed in the hope that it will be useful,              
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of               
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the                 
+ * GNU Lesser General Public License for more details.                          
+ *                                                                            
+ * You should have received a copy of the GNU Lesser General Public License     
+ * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
+ */
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Reflection;
+
+namespace BH.Engine.Reflection
+{
+    public static partial class Create
+    {
+        /***************************************************/
+        /**** Public Methods                            ****/
+        /***************************************************/
+
+        public static Type GenericType(string name, bool silent = false)
+        {
+            if (name == null)
+            {
+                Compute.RecordError("Cannot create a type from a null string.");
+                return null;
+            }
+
+            string[] parts = name.Split('<', '>', ',').Select(x => x.Trim()).ToArray();
+            string[] arguments = parts.Skip(1).Where(x => x.Length > 0).ToArray();
+
+            Type typeDefinition = Type(parts[0] + "`" + arguments.Length);
+            if (typeDefinition == null)
+                return null;
+
+            try
+            {
+                return typeDefinition.MakeGenericType(arguments.Select(x => Type(x)).ToArray());
+            }
+            catch
+            {
+                return null;
+            }
+        }
+
+        /***************************************************/
+    }
+}
